requirements.txt
langchain
langchain-openai
langchain-huggingface
langchain-pinecone
pdfplumber
pandas
python-docx
beautifulsoup4
requests
fastapi
pydantic
python-dotenv
pinecone-client
sentence-transformers
openpyxl
langchain-text-splitters
uvicorn
langchain-huggingface